+++ config/webpack.offscreen.js
@@ -1,9 +1,0 @@
-const PATHS = require("./paths");
-
-const offscreen = {
-  entry: {
-    offscreen: PATHS.src + "/offscreen/index.js",
-  },
-};
-
-module.exports = offscreen;