+++ src/api/message.js
@@ -1,8 +1,0 @@
-export class Message {
-    action;
-    callbackId;
-    param;
-    error;
-    data;
-  }
-