+++ src/common/api/message.js
@@ -1,0 +1,10 @@
+export class Message {
+    action;
+    callbackId;
+    param;
+    error;
+    data;
+    from;
+    to;
+  }
+