+++ src/sidepanel/views/HomeView.vue
@@ -274,8 +274,8 @@
 <script lang="ts" setup>
 import { onMounted, ref, computed, provide, nextTick } from "vue";
 import { useTransition } from "@vueuse/core";
-import { JobApi } from "@/api/index.js";
-import { SearchJobBO } from "@/data/bo/searchJobBO.js";
+import { JobApi } from "../../common/api/index.js";
+import { SearchJobBO } from "../../common/data/bo/searchJobBO.js";
 import dayjs from "dayjs";
 import { utils, writeFileXLSX } from "xlsx";
 import { use } from "echarts/core";
@@ -287,7 +287,7 @@ import {
   LegendComponent,
 } from "echarts/components";
 import VChart from "vue-echarts";
-import { toLine } from "@/utils";
+import { toLine } from "../../common/utils";
 
 use([
   CanvasRenderer,
@@ -551,3 +551,4 @@ const refreshStatistic = async () => {
   height: 400px;
 }
 </style>
+../../common/api/index.js../../common/data/bo/searchJobBO.js../../common/utils/index.js